Move agent impl to shared module to server one
