+++ agents/unison/pom.xml
@@ -27,7 +27,7 @@
         </dependency>
         <dependency>
             <groupId>org.eclipse.che.core</groupId>
-            <artifactId>che-core-api-agent-shared</artifactId>
+            <artifactId>che-core-api-agent</artifactId>
         </dependency>
     </dependencies>
 </project>
