+++ infrastructures/docker/src/test/java/org/eclipse/che/workspace/infrastructure/docker/AgentConfigApplierTest.java
@@ -13,8 +13,8 @@ package org.eclipse.che.workspace.infrastructure.docker;
 import org.eclipse.che.api.agent.server.AgentRegistry;
 import org.eclipse.che.api.agent.server.impl.AgentSorter;
 import org.eclipse.che.api.agent.server.launcher.AgentLauncherFactory;
-import org.eclipse.che.api.agent.shared.model.impl.AgentImpl;
-import org.eclipse.che.api.agent.shared.model.impl.AgentKeyImpl;
+import org.eclipse.che.api.agent.server.model.impl.AgentImpl;
+import org.eclipse.che.api.agent.server.model.impl.AgentKeyImpl;
 import org.eclipse.che.api.core.model.workspace.config.ServerConfig;
 import org.eclipse.che.api.workspace.server.model.impl.MachineConfigImpl;
 import org.eclipse.che.workspace.infrastructure.docker.model.DockerContainerConfig;
