+++ wsagent/agent/pom.xml
@@ -39,7 +39,7 @@
         </dependency -->
         <dependency>
             <groupId>org.eclipse.che.core</groupId>
-            <artifactId>che-core-api-agent-shared</artifactId>
+            <artifactId>che-core-api-agent</artifactId>
         </dependency>
         <!-- dependency>
             <groupId>org.eclipse.che.core</groupId>
